% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ical}
\alias{create_ical}
\title{Create a Outlook calendar invite}
\usage{
create_ical(session, path)
}
\arguments{
\item{session}{dataframe. A single row from  the sessions dataset,
describing a one session instance.}

\item{path}{string. Path to file, where the session should be stored}
}
\description{
Create a Outlook calendar invite
}
