% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_schedule}
\alias{parse_schedule}
\title{Parse R Cafe Schedule}
\usage{
parse_schedule(schedule)
}
\arguments{
\item{schedule}{data frame of R Cafe schedule as provided by \code{retrieve_schedule()}.}
}
\value{
Returns the original data frame with additional columns.
}
\description{
Function to extract additional information from the basic schedule from Google
Sheets to be used for creating calendar events and website descriptions.
}
