% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{upcoming_session}
\alias{upcoming_session}
\title{Extract information on upcoming sessions}
\usage{
upcoming_session(url = "https://delft-rcafe.github.io/home", session_no = 1)
}
\arguments{
\item{url}{string. Give the url for the R Cafe website.}

\item{session_no}{integer. The number of the upcoming session. 1 = first, 2 = second, etc.}
}
\description{
Extract information on upcoming sessions
}
