% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{upcoming_sessions}
\alias{upcoming_sessions}
\title{Extract information on upcoming sessions}
\usage{
upcoming_sessions(n = NULL)
}
\arguments{
\item{n}{integer. Number of sessions to show. Default is to show all upcoming sessions. n = 1 will show the next session.}
}
\description{
Extract information on upcoming sessions
}
