% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_poster}
\alias{update_poster}
\title{Create or update poster}
\usage{
update_poster(session, out_path = "")
}
\arguments{
\item{session}{dataframe. A single row from  the sessions dataset,
describing a one session instance.}

\item{out_path}{string. Path to file, where the session should be stored}
}
\description{
Create or update poster
}
